+++ packages/workbench/src/providers/time-provider.ts
@@ -1,9 +1,10 @@
 import type { Time } from '@use-gpu/core';
 import { makeContext, useContext, useNoContext } from '@use-gpu/live';
 
-export type TimeContextProps = Time;
+export type TimeContextProps = Time & {start: number};
 
 export const TimeContext = makeContext<TimeContextProps>({
+  start: 0,
   timestamp: 0,
   elapsed: 0,
   delta: 0,