+++ packages/workbench/src/shader/index.ts
@@ -1,3 +1,4 @@
 export * from './data-shader';
 export * from './full-screen';
 export * from './texture-shader';
+export * from './value-shader';